+++ mintpy/ifgram_inversion.py
@@ -15,11 +15,11 @@ import time
 import argparse
 import warnings
 import h5py
-import math
 import numpy as np
 from scipy import linalg   # more effieint than numpy.linalg
 from mintpy.objects import ifgramStack, timeseries
 from mintpy.utils import readfile, writefile, ptime, utils as ut
+from mintpy.simulation import decorrelation as decor
 
 
 # key configuration parameter name
@@ -267,115 +267,8 @@ def run_or_skip(inps):
     return flag
 
 
-################################################################################################
-def phase_pdf_ds(L, coherence=None, phi_num=1000, epsilon=1e-3):
-    """Marginal PDF of interferometric phase for distributed scatterers (DS)
-    Eq. 66 (Tough et al., 1995) and Eq. 4.2.23 (Hanssen, 2001)
-    Inputs:
-        L         - int, number of independent looks
-        coherence - 1D np.array for the range of coherence, with value < 1.0 for valid operation
-        phi_num    - int, number of phase sample for the numerical calculation
-    Output:
-        pdf       - 2D np.array, phase pdf in size of (phi_num, len(coherence))
-        coherence - 1D np.array for the range of coherence
-    Example:
-        epsilon = 1e-4
-        coh = np.linspace(0., 1-epsilon, 1000)
-        pdf, coh = phase_pdf_ds(1, coherence=coh)
-    """
-    if coherence is None:
-        coherence = np.linspace(0., 1.-epsilon, 1000)
-    coherence = np.array(coherence, np.float64).reshape(1, -1)
-    phi = np.linspace(-np.pi, np.pi, phi_num, dtype=np.float64).reshape(-1, 1)
-
-    # Phase PDF - Eq. 4.2.32 (Hanssen, 2001)
-    A = np.power((1-np.square(coherence)), L) / (2*np.pi)
-    A = np.tile(A, (phi_num, 1))
-    B = gamma(2*L - 1) / ((gamma(L))**2 * 2**(2*(L-1)))
-
-    beta = np.multiply(np.abs(coherence), np.cos(phi))
-    C = np.divide((2*L - 1) * beta, np.power((1 - np.square(beta)), L+0.5))
-    C = np.multiply(C, (np.pi/2 + np.arcsin(beta)))
-    C += 1 / np.power((1 - np.square(beta)), L)
-
-    sumD = 0
-    if L > 1:
-        for r in range(int(L)-1):
-            D = gamma(L-0.5) / gamma(L-0.5-r)
-            D *= gamma(L-1-r) / gamma(L-1)
-            D *= (1 + (2*r+1)*np.square(beta)) / np.power((1 - np.square(beta)), r+2)
-            sumD += D
-        sumD /= (2*(L-1))
-
-    pdf = B*C + sumD
-    pdf = np.multiply(A, pdf)
-    return pdf, coherence.flatten()
-
-
-def gamma(x):
-    """
-    Gamma function equivalent to scipy.special.gamma(x)
-
-    :param x: float
-    :return: float
-    """
-    # This function replaces scipy.special.gamma(x).
-    # It is needed due to a bug where Dask workers throw an exception in which they cannot
-    # find `scipy.special.gamma(x)` even when it is imported.
-
-    # When the output of the gamma function is undefined, scipy.special.gamma(x) returns float('inf')
-    # whereas math.gamma(x) throws an exception.
-    try:
-        return math.gamma(x)
-    except ValueError:
-        return float('inf')
-
-
-def phase_variance_ds(L,  coherence=None, epsilon=1e-3):
-    """Interferometric phase variance for distributed scatterers (DS)
-    Eq. 2.1.2 (Box et al., 2015) and Eq. 4.2.27 (Hanssen, 2001)
-    Inputs:
-        L         - int, number of independent looks
-        coherence - 1D np.array for the range of coherence, with value < 1.0 for valid operation
-    Output:
-        var       - 1D np.array, phase variance in size of (len(coherence))
-        coherence - 1D np.array for the range of coherence
-    Example:
-        epsilon = 1e-4
-        coh = np.linspace(0., 1-epsilon, 1000)
-        var, coh = phase_variance_ds(1, coherence=coh)
-    """
-    if coherence is None:
-        coherence = np.linspace(0., 1.-epsilon, 1000, dtype=np.float64)
-    phi_num = len(coherence)
-
-    phi = np.linspace(-np.pi, np.pi, phi_num, dtype=np.float64).reshape(-1, 1)
-    phi_step = 2*np.pi/phi_num
-
-    pdf, coherence = phase_pdf_ds(L, coherence=coherence, phi_num=phi_num)
-    var = np.sum(np.multiply(np.square(np.tile(phi, (1, len(coherence)))), pdf)*phi_step, axis=0)
-
-    # assign negative value (when coherence is very close to 1, i.e. 0.999) to the min positive value
-    flag = var <= 0
-    if not np.all(flag):
-        var[flag] = np.nanmin(var[~flag])
-    else:
-        var[flag] = np.finfo(np.float64).eps
-    return var, coherence
-
-
-def phase_variance_ps(L, coherence=None, epsilon=1e-3):
-    """the Cramer-Rao bound (CRB) of phase variance
-    Given by Eq. 25 (Rodriguez and Martin, 1992)and Eq 4.2.32 (Hanssen, 2001)
-    Valid when coherence is close to 1.
-    """
-    if coherence is None:
-        coherence = np.linspace(0.9, 1.-epsilon, 1000, dtype=np.float64)
-    var = (1 - coherence**2) / (2 * int(L) * coherence**2)
-    return var, coherence
-
-
-def coherence2phase_variance_ds(coherence, L=32, epsilon=1e-3, print_msg=False):
+#################################### Weight Functions #####################################
+def coherence2phase_variance(coherence, L=32, epsilon=1e-3, print_msg=False):
     """Convert coherence to phase variance based on DS phase PDF (Tough et al., 1995)"""
     lineStr = '    number of looks L={}'.format(L)
     if L > 80:
@@ -397,7 +290,7 @@ def coherence2phase_variance_ds(coherence, L=32, epsilon=1e-3, print_msg=False):
     coherence[coherence > coh_max] = coh_max
     coherence_idx = np.array((coherence - coh_min) / coh_step, np.int16)
 
-    var_lut = phase_variance_ds(int(L), coh_lut)[0]
+    var_lut = decor.phase_variance_ds(int(L), coh_lut)[0]
     variance = var_lut[coherence_idx]
     return variance
 
@@ -411,6 +304,42 @@ def coherence2fisher_info_index(data, L=32, epsilon=1e-3):
     return data
 
 
+def coherence2weight(coh_data, weight_func='var', L=20, epsilon=5e-2, print_msg=True):
+    coh_data[np.isnan(coh_data)] = epsilon
+    coh_data[coh_data < epsilon] = epsilon
+    coh_data = np.array(coh_data, np.float64)
+
+    # Calculate Weight matrix
+    weight_func = weight_func.lower()
+    if 'var' in weight_func:
+        if print_msg:
+            print('convert coherence to weight using inverse of phase variance')
+            print('    with phase PDF for distributed scatterers from Tough et al. (1995)')
+        weight = 1.0 / coherence2phase_variance(coh_data, L, print_msg=print_msg)
+
+    elif any(i in weight_func for i in ['coh', 'lin']):
+        if print_msg:
+            print('use coherence as weight directly (Perissin & Wang, 2012; Tong et al., 2016)')
+        weight = coh_data
+
+    elif any(i in weight_func for i in ['fim', 'fisher']):
+        if print_msg:
+            print('convert coherence to weight using Fisher Information Index (Seymour & Cumming, 1994)')
+        weight = coherence2fisher_info_index(coh_data, L)
+
+    elif weight_func in ['no', 'sbas', 'uniform']:
+        weight = None
+
+    else:
+        raise Exception('Un-recognized weight function: %s' % weight_func)
+
+    if weight is not None:
+        weight = np.array(weight, np.float32)
+    del coh_data
+    return weight
+
+
+################################# Time-series Estimator ###################################
 def estimate_timeseries(A, B, tbase_diff, ifgram, weight_sqrt=None, min_norm_velocity=True,
                         rcond=1e-5, min_redundancy=1.):
     """Estimate time-series from a stack/network of interferograms with
@@ -525,7 +454,7 @@ def estimate_timeseries(A, B, tbase_diff, ifgram, weight_sqrt=None, min_norm_vel
     return ts, temp_coh, num_inv_ifg
 
 
-###########################################################################################
+###################################### File IO ############################################
 def write2hdf5_file(ifgram_file, metadata, ts, temp_coh, num_inv_ifg=None,
                     suffix='', inps=None):
     stack_obj = ifgramStack(ifgram_file)
@@ -784,41 +713,6 @@ def read_coherence(stack_obj, box, dropIfgram=True, print_msg=True):
                               print_msg=False).reshape(num_ifgram, -1)
     coh_data[np.isnan(coh_data)] = 0.
     return coh_data
-
-
-def coherence2weight(coh_data, weight_func='var', L=20, epsilon=5e-2, print_msg=True):
-    coh_data[np.isnan(coh_data)] = epsilon
-    coh_data[coh_data < epsilon] = epsilon
-    coh_data = np.array(coh_data, np.float64)
-
-    # Calculate Weight matrix
-    weight_func = weight_func.lower()
-    if 'var' in weight_func:
-        if print_msg:
-            print('convert coherence to weight using inverse of phase variance')
-            print('    with phase PDF for distributed scatterers from Tough et al. (1995)')
-        weight = 1.0 / coherence2phase_variance_ds(coh_data, L, print_msg=print_msg)
-
-    elif any(i in weight_func for i in ['coh', 'lin']):
-        if print_msg:
-            print('use coherence as weight directly (Perissin & Wang, 2012; Tong et al., 2016)')
-        weight = coh_data
-
-    elif any(i in weight_func for i in ['fim', 'fisher']):
-        if print_msg:
-            print('convert coherence to weight using Fisher Information Index (Seymour & Cumming, 1994)')
-        weight = coherence2fisher_info_index(coh_data, L)
-
-    elif weight_func in ['no', 'sbas', 'uniform']:
-        weight = None
-
-    else:
-        raise Exception('Un-recognized weight function: %s' % weight_func)
-
-    if weight is not None:
-        weight = np.array(weight, np.float32)
-    del coh_data
-    return weight
 
 
 def ifgram_inversion_patch(ifgram_file, box=None, ref_phase=None, unwDatasetName='unwrapPhase',
