+++ mintpy/simulation/README.md
@@ -1,0 +1,7 @@
+## Simulation Toolbox for Interferometric Phase
+
+This module collects some toolbox to simulate different components of interferometric phases, such as the decorrelation noise, tropospheric turbulence and surface deformation, etc. 
+
+This toolbox is still in **alpha** stage, which means they will change quite frequently. 
+
+There are no comprehensive document on this part yet. If you want to give them a try, read the detailed function usage in the code.
