+++ mintpy/simulation/__init__.py
@@ -2,6 +2,3 @@
 Toolbox for InSAR time-series analysis simulation.
 """
 
-
-from .simulation import *
-
